chore: remove nested atom
