+++ src/store/atom.ts
@@ -1,0 +1,6 @@
+import { atom } from "recoil";
+
+export const likedBeerState = atom({
+  key: "likedBeerState",
+  default: null,
+});
