+++ src/store/selector.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
 import { selector } from "recoil";
-import { BeerResultType } from "../../../../interface/types";
+import { BeerResultType } from "../../interface/types";
 
 export const popularBeerState = selector<BeerResultType>({
   key: "popularBeerStateReadOnly",
@@ -10,5 +10,13 @@ export const popularBeerState = selector<BeerResultType>({
       throw res.error;
     }
     return res;
+  },
+});
+
+export const likedBeerQuery = selector({
+  key: "likedBeerQuery",
+  get: async () => {
+    const response = await fetch(`/api/v1/beers/1`).then((res) => res.json());
+    return response;
   },
 });
